Optional axis argument for cumulative functions (#2664)
* Make cumreduction's axis argument optional

Have `cumreduction`'s `axis` argument default to `None`. If it is
`None`, simply flatten the array and set `axis` to `0`. This is the
smallest change to support `axis` as `None`. Also it results in
equivalent behavior to what NumPy would do otherwise.

* Make `axis` optional for `cumsum` and `cumprod`

This is matches what NumPy expects for these parameters. Also the
handling of them is the same as well thanks to the change made to
`cumreduction`.

* Test cumulative reductions

Provide some tests of cumulative reductions on slightly more interesting
data and with different specifications of `axis`.