Add ptp (#2691)
* Add ptp

Implements `ptp` for Dask Arrays. Basically just the difference between
the `max` and `min` with an optional `axis` argument. Skipped providing
`out` even though NumPy provides it. Not sure that it makes much sense
to have `out` for Dask Arrays or that it is of any value.

* Export ptp in Dask Array's public API

* Document ptp in Dask Array API docs

* Test ptp

Provide some test cases for the Dask Array `ptp` implementation with
different `axis` arguments.